Relocate Reservation component and update App test
